+++ web/app/components/core/application.component.jsx
@@ -4,14 +4,12 @@ var RouteHandler = require('react-router').RouteHandler;
 
 var Application = React.createClass({
   render: function() {
-    /* jshint ignore:start */
     return (
       <div className="application">
         <Header />
         <RouteHandler />
       </div>
     );
-    /* jshint ignore:end */
   }
 });