+++ web/app/components/core/header.component.jsx
@@ -24,7 +24,6 @@ var Header = React.createClass({
   },
 
   render: function() {
-    /* jshint ignore:start */
     return (
       <header>
         <ul>
@@ -36,7 +35,6 @@ var Header = React.createClass({
         </ul>
       </header>
     );
-    /* jshint ignore:end */
   }
 });