+++ web/app/components/desktop/prevent-double-click.component.jsx
@@ -29,7 +29,6 @@ var PreventDoubleClick = React.createClass({
   },
 
   render: function() {
-    /* jshint ignore:start */
     return (
       <div>
         <h1 id="test" className="test">{window.i18n['desktop/prevent-double-click'].header()}</h1>
@@ -39,7 +38,6 @@ var PreventDoubleClick = React.createClass({
         </div>
       </div>
     );
-    /* jshint ignore:end */
   },
 
   _callback: function() {