+++ web/app/components/desktop/with-resolves.component.jsx
@@ -23,11 +23,9 @@ var WithResolves = React.createClass({
     });
   },
   render: function() {
-    /* jshint ignore:start */
     return (
       <h1 id="test" className="test">{window.i18n['desktop/with-resolves'].header()}</h1>
     );
-    /* jshint ignore:end */
   }
 });